pyproject.toml
[tool.poetry]
name = "suaischeduleapi2"
version = "0.1.0"
description = ""
authors = ["triks <triks0001000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.109.1"}
django-environ = "^0.11.2"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
